backend/requirements.txt
Flask
Flask-Security
Flask-SQLAlchemy
Flask-Cors
Flask-Caching
pytz
celery
redis
